requirements.txt
Flask==3.0.3
Flask_Parameter_Validation==2.3.1
openai==1.30.1
pdf2image==1.17.0
python-dotenv==1.0.1
Werkzeug==3.0.3
gunicorn==22.0.0
wheel==0.43.0
boto3==1.34.122